Move unified symbol APIs into their own file
