+++ Sources/SwiftDocC/Semantics/Symbol/Symbol.swift
@@ -372,94 +372,6 @@ extension Symbol {
     }
 }
 
-extension UnifiedSymbolGraph.Symbol {
-    var defaultSelector: UnifiedSymbolGraph.Selector? {
-        if let mainSelector = self.mainGraphSelectors.first, mainSelector.interfaceLanguage == "swift" {
-            return mainSelector
-        }
-
-        let defaultSelector = UnifiedSymbolGraph.Selector(interfaceLanguage: "swift", platform: nil)
-
-        if self.pathComponents.keys.contains(defaultSelector) {
-            return defaultSelector
-        } else if let firstSwiftSelector = self.pathComponents.keys.first(where: { $0.interfaceLanguage == "swift" }) {
-            return firstSwiftSelector
-        } else if FeatureFlags.current.isExperimentalObjectiveCSupportEnabled {
-            return mainGraphSelectors.first ?? pathComponents.keys.first
-        } else {
-            return nil
-        }
-    }
-
-    func symbol(forSelector selector: UnifiedSymbolGraph.Selector?) -> SymbolGraph.Symbol? {
-        guard let selector = selector,
-              let kind = self.kind[selector],
-              let pathComponents = self.pathComponents[selector],
-              let names = self.names[selector],
-              let accessLevel = self.accessLevel[selector],
-              let mixins = self.mixins[selector] else {
-            return nil
-        }
-
-        return SymbolGraph.Symbol(
-            identifier: SymbolGraph.Symbol.Identifier(
-                precise: self.uniqueIdentifier,
-                interfaceLanguage: selector.interfaceLanguage),
-            names: names,
-            pathComponents: pathComponents,
-            docComment: self.docComment[selector],
-            accessLevel: accessLevel,
-            kind: kind,
-            mixins: mixins
-        )
-    }
-
-    var defaultSymbol: SymbolGraph.Symbol? {
-        symbol(forSelector: defaultSelector)
-    }
-    
-    /// Returns the primary symbol to use as documentation source.
-    var documentedSymbol: SymbolGraph.Symbol? {
-        guard FeatureFlags.current.isExperimentalObjectiveCSupportEnabled else {
-            return defaultSymbol
-        }
-        
-        return symbol(forSelector: documentedSymbolSelector)
-    }
-    
-    /// Returns the primary symbol selector to use as documentation source.
-    var documentedSymbolSelector: UnifiedSymbolGraph.Selector? {
-        guard FeatureFlags.current.isExperimentalObjectiveCSupportEnabled else {
-            return defaultSelector
-        }
-        
-        // We'll prioritize the first documented 'swift' symbol, if we have
-        // one.
-        return docComment.keys.first { selector  in
-            return selector.interfaceLanguage == "swift"
-        } ?? docComment.keys.first
-    }
-
-    func identifier(forLanguage interfaceLanguage: String) -> SymbolGraph.Symbol.Identifier {
-        return SymbolGraph.Symbol.Identifier(
-            precise: self.uniqueIdentifier,
-            interfaceLanguage: interfaceLanguage
-        )
-    }
-
-    var defaultIdentifier: SymbolGraph.Symbol.Identifier {
-        guard FeatureFlags.current.isExperimentalObjectiveCSupportEnabled else {
-            return identifier(forLanguage: "swift")
-        }
-        
-        if let defaultInterfaceLanguage = defaultSelector?.interfaceLanguage {
-            return identifier(forLanguage: defaultInterfaceLanguage)
-        } else {
-            return identifier(forLanguage: "swift")
-        }
-    }
-}
-
 extension Dictionary where Key == String, Value == Mixin {
     func getValueIfPresent<T>(for mixinType: T.Type) -> T? where T: Mixin {
         return self[mixinType.mixinKey] as? T
